[수정] amazon carousel  중복사항 수정
