+++ .vscode/launch.json
@@ -1,0 +1,15 @@
+{
+  // IntelliSense를 사용하여 가능한 특성에 대해 알아보세요.
+  // 기존 특성에 대한 설명을 보려면 가리킵니다.
+  // 자세한 내용을 보려면 https://go.microsoft.com/fwlink/?linkid=830387을(를) 방문하세요.
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "프로그램 시작",
+      "skipFiles": ["<node_internals>/**"],
+      "program": "${workspaceFolder}\\main\\JS\\test.js"
+    }
+  ]
+}
