+++ main/amazon.html
@@ -20,6 +20,6 @@
         </div>
       </div>
     </div>
-    <script src="./JS/mainUI.js" type="module"></script>
+    <script src="./JS/fetch.js" type="module"></script>
   </body>
 </html>
